requirements.txt
flask
pandas
openpyxl
python-docx
PyMuPDF
nltk
sentence-transformers
scikit-learn
werkzeug
scikit-learn
nltk
pymupdf
pandas
python-docx
openpyxl
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-grpc
pymupdf
pandas
python-docx
openpyxl
pytesseract
pillow
poppler-utils
pdf2image